refactor: 카카오부분 api,viewmodel,repository 분리
