+++ data/src/main/kotlin/com/sooum/data/di/NetworkModule.kt
@@ -4,6 +4,7 @@ import com.sooum.data.network.NullOnEmptyConverterFactory
 import com.sooum.data.network.auth.AuthApi
 import com.sooum.data.network.comment.CommentApi
 import com.sooum.data.network.friend.FriendApi
+import com.sooum.data.network.kakao.KakaoApi
 import com.sooum.data.network.meet.MeetApi
 import com.sooum.data.network.place.PlaceApi
 import com.sooum.data.network.schedule.ScheduleApi
@@ -102,5 +103,13 @@ object NetworkModule {
         @WhereRetrofit retrofit: Retrofit
     ): AuthApi {
         return retrofit.create(AuthApi::class.java)
+    }
+
+    @Provides
+    @Singleton
+    fun provideKakaoApi(
+        @WhereRetrofit retrofit: Retrofit
+    ): KakaoApi {
+        return retrofit.create(KakaoApi::class.java)
     }
 }
