+++ data/src/main/kotlin/com/sooum/data/di/RepositoryModule.kt
@@ -2,11 +2,13 @@ package com.sooum.data.di
 
 import com.sooum.data.repository.AuthRepositoryImpl
 import com.sooum.data.repository.FriendRepositoryImpl
+import com.sooum.data.repository.KakaoRepositoryImpl
 import com.sooum.data.repository.MeetDetailCommentRepositoryImpl
 import com.sooum.data.repository.MeetDetailPlaceRepositoryImpl
 import com.sooum.data.repository.MeetDetailRepositoryImpl
 import com.sooum.domain.repository.AuthRepository
 import com.sooum.domain.repository.FriendRepository
+import com.sooum.domain.repository.KakaoRepository
 import com.sooum.domain.repository.MeetDetailCommentRepository
 import com.sooum.domain.repository.MeetDetailPlaceRepository
 import com.sooum.domain.repository.MeetDetailRepository
@@ -50,4 +52,10 @@ abstract class RepositoryModule {
     abstract fun bindAuthRepository(
         impl: AuthRepositoryImpl
     ): AuthRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindKakaoRepository(
+        impl: KakaoRepositoryImpl
+    ): KakaoRepository
 }
