+++ domain/src/main/kotlin/com/sooum/domain/usecase/kakao/KakaoSignUpUseCase.kt
@@ -1,18 +1,18 @@
-package com.sooum.domain.usecase.auth
+package com.sooum.domain.usecase.kakao
 
 import com.sooum.domain.model.ApiResult
 import com.sooum.domain.model.KakaoSignUpResult
-import com.sooum.domain.repository.AuthRepository
+import com.sooum.domain.repository.KakaoRepository
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.Flow
 
 class KakaoSignUpUseCase @Inject constructor(
-    private val authRepository: AuthRepository
+    private val kakaoRepository: KakaoRepository
 ) {
     suspend operator fun invoke(
         accessToken: String,
         refreshToken: String
     ): Flow<ApiResult<KakaoSignUpResult>> {
-        return authRepository.kakaoLogin(accessToken, refreshToken)
+        return kakaoRepository.kakaoLogin(accessToken, refreshToken)
     }
 }
