+++ domain/src/main/kotlin/com/sooum/domain/usecase/kakao/NickNameUpdateUseCase.kt
@@ -1,14 +1,14 @@
-package com.sooum.domain.usecase.auth
+package com.sooum.domain.usecase.kakao
 
 import com.sooum.domain.model.ApiResult
-import com.sooum.domain.repository.AuthRepository
+import com.sooum.domain.repository.KakaoRepository
 import jakarta.inject.Inject
 import kotlinx.coroutines.flow.Flow
 
 class NickNameUpdateUseCase @Inject constructor(
-    private val authRepository: AuthRepository
+    private val kakaoRepository: KakaoRepository
 ) {
     suspend operator fun invoke(userId: Int, nickName: String): Flow<ApiResult<Unit>>{
-        return authRepository.putNickName(userId, nickName)
+        return kakaoRepository.putNickName(userId, nickName)
     }
 }
