+++ presentation/src/main/kotlin/com/sooum/where_android/viewmodel/AuthViewModel.kt
@@ -1,18 +1,11 @@
 package com.sooum.where_android.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import androidx.lifecycle.viewmodel.compose.viewModel
 import com.sooum.domain.model.ApiResult
-import com.sooum.domain.model.EmailVerifyResult
-import com.sooum.domain.model.KakaoSignUpResult
 import com.sooum.domain.model.SignUpResult
 import com.sooum.domain.usecase.auth.EmailVerifyUseCase
-import com.sooum.domain.usecase.auth.KakaoSignUpUseCase
 import com.sooum.domain.usecase.auth.LoginUseCase
-import com.sooum.domain.usecase.auth.NickNameUpdateUseCase
-import com.sooum.domain.usecase.auth.ProfileUpdateUseCase
 import com.sooum.domain.usecase.auth.RequestEmailAuthUseCase
 import com.sooum.domain.usecase.auth.SignUpUseCase
 import com.sooum.domain.usecase.auth.ValidatePasswordUseCase
@@ -22,25 +15,18 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import java.io.File
 
 @HiltViewModel
 class AuthViewModel @Inject constructor(
     private val validatePasswordUseCase: ValidatePasswordUseCase,
     private val loginUseCase: LoginUseCase,
     private val signUpUseCase: SignUpUseCase,
-    private val kakaoSignUpUseCase: KakaoSignUpUseCase,
-    private val nickNameUpdateUseCase: NickNameUpdateUseCase,
-    private val profileUpdateUseCase: ProfileUpdateUseCase,
     private val getRequestEmailAuthUseCase: RequestEmailAuthUseCase,
     private val emailVerifyUseCase: EmailVerifyUseCase
 ) : ViewModel(){
 
     private val _loginState = MutableStateFlow<ApiResult<Any>>(ApiResult.SuccessEmpty)
     val loginState: StateFlow<ApiResult<Any>> = _loginState.asStateFlow()
-
-    private val _updateNicknameState = MutableStateFlow<ApiResult<Unit>>(ApiResult.SuccessEmpty)
-    val updateNicknameState: StateFlow<ApiResult<Unit>> = _updateNicknameState
   
     private val _emailRequestState = MutableStateFlow<ApiResult<Unit>>(ApiResult.SuccessEmpty)
     val emailRequestState: StateFlow<ApiResult<Unit>> = _emailRequestState.asStateFlow()
@@ -51,8 +37,6 @@ class AuthViewModel @Inject constructor(
     private val _signUpState = MutableStateFlow<ApiResult<SignUpResult>>(ApiResult.SuccessEmpty)
     val signUpState: StateFlow<ApiResult<SignUpResult>> = _signUpState
 
-    private val _kakaoSignUpState = MutableStateFlow<ApiResult<KakaoSignUpResult>>(ApiResult.SuccessEmpty)
-    val kakaoSignUpState: StateFlow<ApiResult<KakaoSignUpResult>> = _kakaoSignUpState
 
     private val _password = MutableStateFlow("")
     val password: StateFlow<String> = _password.asStateFlow()
@@ -106,19 +90,6 @@ class AuthViewModel @Inject constructor(
     }
 
     /**
-     * 카카오 회원가입 기능
-     */
-    fun kakaoLogin(authorization: String, refreshToken: String) {
-        viewModelScope.launch {
-           kakaoSignUpUseCase(
-               authorization ,refreshToken
-           ).collect{ result ->
-               _kakaoSignUpState.value = result
-           }
-        }
-    }
-
-    /**
      * 회원가입 기능
      */
     fun signUp(
@@ -131,32 +102,6 @@ class AuthViewModel @Inject constructor(
                 profileImage = _profileImage.value
             ).collect { result ->
                 _signUpState.value = result
-            }
-        }
-    }
-
-    /**
-     * 닉네임 수정 기능
-     */
-    fun putNickName(userId: Int, nickName: String){
-        viewModelScope.launch {
-            nickNameUpdateUseCase(
-                userId, nickName
-            ).collect{ result->
-                _updateNicknameState.value = result
-            }
-        }
-    }
-
-    /**
-     * 프로필 업데이트 기능
-     */
-    fun updateProfile(userId: Int,imageFile: File?){
-        viewModelScope.launch {
-            profileUpdateUseCase(
-                userId, imageFile
-            ).collect{ result ->
-              
             }
         }
     }
